chore(linter): remove dependency on @nrwl/workspace
