+++ packages/angular/src/generators/utils/storybook-ast/storybook-inputs.ts
@@ -1,8 +1,6 @@
 import type { Tree } from '@nrwl/devkit';
-import {
-  findNodes,
-  getSourceNodes,
-} from '@nrwl/workspace/src/utilities/typescript';
+import { findNodes } from 'nx/src/utils/typescript';
+import { getSourceNodes } from '@nrwl/workspace/src/utilities/typescript';
 import type { PropertyDeclaration } from 'typescript';
 import { SyntaxKind } from 'typescript';
 import { getTsSourceFile } from '../../../utils/nx-devkit/ast-utils';
