+++ packages/eslint-plugin-nx/src/utils/graph-utils.spec.ts
