+++ packages/eslint-plugin-nx/src/utils/graph-utils.ts
