+++ packages/nx/src/utils/typescript.ts
@@ -2,6 +2,7 @@ import { workspaceRoot } from './workspace-root';
 import { existsSync } from 'fs';
 import { dirname, join } from 'path';
 import type * as ts from 'typescript';
+import type { Node, SyntaxKind } from 'typescript';
 
 const normalizedAppRoot = workspaceRoot.replace(/\\/g, '/');
 
@@ -103,4 +104,39 @@ export function getRootTsConfigPath(): string | null {
   const tsConfigFileName = getRootTsConfigFileName();
 
   return tsConfigFileName ? join(workspaceRoot, tsConfigFileName) : null;
+}
+
+export function findNodes(
+  node: Node,
+  kind: SyntaxKind | SyntaxKind[],
+  max = Infinity
+): Node[] {
+  if (!node || max == 0) {
+    return [];
+  }
+
+  const arr: Node[] = [];
+  const hasMatch = Array.isArray(kind)
+    ? kind.includes(node.kind)
+    : node.kind === kind;
+  if (hasMatch) {
+    arr.push(node);
+    max--;
+  }
+  if (max > 0) {
+    for (const child of node.getChildren()) {
+      findNodes(child, kind, max).forEach((node) => {
+        if (max > 0) {
+          arr.push(node);
+        }
+        max--;
+      });
+
+      if (max <= 0) {
+        break;
+      }
+    }
+  }
+
+  return arr;
 }
