+++ packages/workspace/src/generators/move/lib/update-imports.ts
@@ -12,7 +12,7 @@ import {
 import { getImportPath } from 'nx/src/utils/path';
 import * as ts from 'typescript';
 import { getRootTsConfigPathInTree } from '../../../utilities/typescript';
-import { findNodes } from '../../../utilities/typescript/find-nodes';
+import { findNodes } from 'nx/src/utils/typescript';
 import { NormalizedSchema } from '../schema';
 import { normalizeSlashes } from './utils';
