+++ packages/workspace/src/utils/ast-utils.ts
@@ -28,7 +28,7 @@ import { getWorkspacePath } from './cli-config-utils';
 import { extname, join, normalize, Path } from '@angular-devkit/core';
 import type { NxJsonConfiguration, ProjectsConfigurations } from '@nrwl/devkit';
 import { addInstallTask } from './rules/add-install-task';
-import { findNodes } from '../utilities/typescript/find-nodes';
+import { findNodes } from 'nx/src/utils/typescript';
 import { getSourceNodes } from '../utilities/typescript/get-source-nodes';
 
 function nodesByPosition(first: ts.Node, second: ts.Node): number {
@@ -71,7 +71,7 @@ export function sortObjectByKeys(obj: unknown) {
     }, {});
 }
 
-export { findNodes } from '../utilities/typescript/find-nodes';
+export { findNodes } from '../utilities/typescript/find-nodes'; // TODO(v16): remove this
 export { getSourceNodes } from '../utilities/typescript/get-source-nodes';
 
 export interface Change {
